backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
deepface==0.0.93
opencv-python==4.9.0.80
mediapipe==0.10.14
numpy==1.26.4
Pillow==10.3.0
pdfplumber==0.11.1
openpyxl==3.1.4
pandas==2.2.2
tf-keras==2.16.0
scipy==1.13.0
